+++ src/client/src/components/DraggableSidebarItem/messages.ts
@@ -1,0 +1,18 @@
+
+import { defineMessages } from "react-intl";
+
+const messages = defineMessages({
+  changeItemName: {
+    id: "draggableSidebarItem.changeItemName",
+    defaultMessage: "Change Item Name"
+  },
+  pageNameMaxLength: {
+    id: "draggableSidebarItem.pageNameMaxLength",
+    defaultMessage: "Page name must be under {maxLength} characters long. "
+  },
+  deleteItem: {
+    id: "draggableSidebarItem.deleteItem",
+    defaultMessage: "Delete item"
+  }
+});
+export default messages;
