+++ src/client/src/components/OutputPath/messages.ts
@@ -1,0 +1,18 @@
+
+import { defineMessages } from "react-intl";
+
+const messages = defineMessages({
+  outputPath: {
+    id: "outputPath.outputPath",
+    defaultMessage: "Output Path"
+  },
+  ariaOutputPathLabel: {
+    id: "outputPath.ariaOutputPath",
+    defaultMessage: "Input for Output Path"
+  },
+  browseButtonLabel: {
+    id: "outputPath.browseButtonLabel",
+    defaultMessage: "Browse"
+  }
+});
+export default messages;
