+++ src/client/src/components/SelectableCard/messages.ts
@@ -1,0 +1,19 @@
+
+import { defineMessages } from "react-intl";
+
+const messages = defineMessages({
+    learnMore: {
+      id: "selectableCard.details",
+      defaultMessage: "Learn more"
+    },
+    preview: {
+      id: "selectableCard.preview",
+      defaultMessage: "Preview"
+    },
+    pageCount: {
+      id: "selectableCard.pageCount",
+      defaultMessage: "{number} {page}"
+    }
+  });
+
+export default messages;
