+++ src/client/src/containers/Details/messages.ts
@@ -1,0 +1,17 @@
+import { defineMessages } from "react-intl";
+
+const messages = defineMessages({
+  none: {
+    id: "details.none",
+    defaultMessage: "None"
+  },
+  redirectLabel: {
+    id: "licenses.redirectLinkLabel",
+    defaultMessage: "{redirectLabel}"
+  },
+  genericRedirect: {
+    id: "redirectLabel.generic",
+    defaultMessage: "here"
+  }
+});
+export default messages;
