+++ src/client/src/containers/Footer/messages.ts
@@ -1,0 +1,9 @@
+import { defineMessages } from "react-intl";
+
+const messages = defineMessages({
+  navAriaLabel: {
+    id: "footer.navAriaLabel",
+    defaultMessage: "Navigate between pages and create project"
+  }
+});
+export default messages;
