+++ src/client/src/containers/ProjectNameAndOutput/messages.ts
@@ -1,0 +1,17 @@
+import { defineMessages } from "react-intl";
+
+const messages = defineMessages({
+  projectNameTitle: {
+    id: "projectName.projectNameTitle",
+    defaultMessage: "Project Name"
+  },
+  ariaProjectNameLabel: {
+    id: "projectName.ariaProjectName",
+    defaultMessage: "Project Name Input"
+  },
+  outputPathTitle: {
+    id: "projectName.outputPathTitle",
+    defaultMessage: "Save To"
+  }
+});
+export default messages;
