+++ src/client/src/containers/RedirectModal/index.tsx
@@ -15,6 +15,7 @@ import { AppState } from "../../reducers";
 import { Dispatch } from "redux";
 import RootAction from "../../actions/ActionType";
 import keyUpHandler from "../../utils/keyUpHandler";
+import messages from "./messages";
 
 export interface IRedirectModalData {
   redirectLink: string;
@@ -33,34 +34,6 @@ interface IDispatchProps {
 }
 
 type Props = IStateProps & IDispatchProps & InjectedIntlProps;
-
-const messages = defineMessages({
-  beingRedirected: {
-    id: "redirectModal.beingRedirected",
-    defaultMessage: "You are being redirected"
-  },
-  thirdPartyWebsite: {
-    id: "redirectModal.thirdPartyWebsite",
-    defaultMessage:
-      "You will be taken to {thirdPartyWebsite} which is a non-Microsoft service."
-  },
-  noneThirdPartyWebsite: {
-    id: "redirectModal.noneThirdPartyWebsite",
-    defaultMessage: "The link will take you to {noneThirdPartyWebsite}."
-  },
-  toContinue: {
-    id: "redirectModal.toContinue",
-    defaultMessage: "To continue, press ok."
-  },
-  privacyStatement: {
-    id: "redirectModal.privacyStatement",
-    defaultMessage: "Privacy Statement"
-  },
-  OK: {
-    id: "redirectModal.OK",
-    defaultMessage: "OK"
-  }
-});
 
 const RedirectModal = (props: Props) => {
   const { privacyData, intl } = props;
