+++ src/client/src/containers/ReviewAndGenerate/messages.ts
@@ -1,0 +1,22 @@
+import { defineMessages } from "react-intl";
+
+const messages = defineMessages({
+  viewLicenses: {
+    id: "licenses.viewLicenses",
+    defaultMessage: "View Licenses"
+  },
+  launchYourProject: {
+    id: "instructionHeading.launchYourProject",
+    defaultMessage: "Launch Your Project"
+  },
+  almostDone: {
+    id: "context.almostDone",
+    defaultMessage:
+      "You're almost done - review your project details on the right and make any necessary adjustments!"
+  },
+  quickStart: {
+    id: "context.quickStart",
+    defaultMessage: `Woo-hoo! You skipped to the end where we'll get you started quickly. Just click "Create Project".`
+  }
+});
+export default messages;
