+++ src/client/src/containers/SelectBackendFramework/messages.ts
@@ -1,0 +1,9 @@
+import { defineMessages } from "react-intl";
+
+const messages = defineMessages({
+  selectBackendFramework: {
+    id: "selectBackendFramework.selectBackendFramework",
+    defaultMessage: "Select a back-end framework"
+  }
+});
+export default messages;
