+++ src/client/src/containers/SelectFrontEndFramework/messages.ts
@@ -1,0 +1,9 @@
+import { defineMessages } from "react-intl";
+
+const messages = defineMessages({
+  selectFrontendFramework: {
+    id: "selectFrontendFramework.selectFrontendFramework",
+    defaultMessage: "Select a front-end framework"
+  }
+});
+export default messages;
