+++ src/client/src/containers/SelectOption/index.tsx
@@ -22,25 +22,7 @@ import { ISelectedPages, mapStateSelectedPages } from "../../selectors/wizardSel
 
 import { InjectedIntl, defineMessages, injectIntl } from "react-intl";
 import { inferItemName} from "../../utils/infer/itemName";
-
-const messages = defineMessages({
-  limitedPages: {
-    id: "pages.limitedPagesMessage",
-    defaultMessage: "You can select up to 20 pages"
-  },
-  overlimitPages: {
-    id: "pages.overlimitPagesMessage",
-    defaultMessage: "You cannot add more than 20 pages to the project"
-  },
-  noPageGeneration: {
-    id: "pages.noPageGeneration",
-    defaultMessage: "At least 1 page must be selected"
-  },
-  iconAltMessage: {
-    id: "pages.maxPagesText",
-    defaultMessage: "Notification"
-  }
-});
+import messages from "./messages";
 
 interface ICount {
   [key: string]: number;
