+++ src/client/src/containers/SelectWebApp/messages.ts
@@ -1,0 +1,9 @@
+import { defineMessages } from "react-intl";
+
+const messages = defineMessages({
+  webAppTitleQuestion: {
+    id: "selectPages.webAppTitleQuestion",
+    defaultMessage: "Select a project type."
+  }
+});
+export default messages;
