+++ src/client/src/containers/SortablePageList/index.tsx
@@ -30,6 +30,7 @@ import { IValidations } from "../../reducers/wizardSelectionReducers/setValidati
 import {
   getValidations
 } from "../../selectors/wizardSelectionSelector/wizardSelectionSelector";
+import messages from "./messages";
 
 interface ISortablePageListProps {
   selectedPages: any[];
@@ -55,25 +56,6 @@ type Props = ISortablePageListProps &
   ISortableDispatchProps &
   IStateProps &
   IIntlProps;
-
-const messages = defineMessages({
-  duplicateName: {
-    id: "sortablePageList.duplicateName",
-    defaultMessage: "Page name has to be unique"
-  },
-  hide: {
-    id: "sortablePageList.hide",
-    defaultMessage: "Hide"
-  },
-  show: {
-    id: "sortablePageList.show",
-    defaultMessage: "Show"
-  },
-  pages: {
-    id: "sortablePageList.pages",
-    defaultMessage: "Pages"
-  }
-});
 
 const SortablePageList = (props: Props) => {
   const {
