+++ src/client/src/containers/SortablePageList/messages.ts
@@ -1,0 +1,21 @@
+import { defineMessages } from "react-intl";
+
+const messages = defineMessages({
+  duplicateName: {
+    id: "sortablePageList.duplicateName",
+    defaultMessage: "Page name has to be unique"
+  },
+  hide: {
+    id: "sortablePageList.hide",
+    defaultMessage: "Hide"
+  },
+  show: {
+    id: "sortablePageList.show",
+    defaultMessage: "Show"
+  },
+  pages: {
+    id: "sortablePageList.pages",
+    defaultMessage: "Pages"
+  }
+});
+export default messages;
