+++ src/client/src/containers/SummarySection/messages.ts
@@ -1,0 +1,9 @@
+import { defineMessages } from "react-intl";
+
+const messages = defineMessages({
+  duplicateFunctionName: {
+    id: "summarySection.duplicateName",
+    defaultMessage: "Function name has to be unique"
+  }
+});
+export default messages;
