+++ src/client/src/containers/ViewLicensesModal/messages.ts
@@ -1,0 +1,13 @@
+import { defineMessages } from "react-intl";
+
+const messages = defineMessages({
+  closeModalLabel: {
+    id: "viewLicensesModal.closeModalLabel",
+    defaultMessage: "Close"
+  },
+  licenses: {
+    id: "licenses.licenses",
+    defaultMessage: "Licenses"
+  }
+});
+export default messages;
