--- src/client/src/messages/comunMessages.ts
@@ -1,9 +1,0 @@
-
-import { defineMessages } from "react-intl";
-
-export const comunMessages = defineMessages({
-  default:{
-    id: "default",
-    defaultMessage: "default"
-  }
-});
