+++ src/client/src/selectors/wizardSelectionSelector/messages.ts
@@ -1,0 +1,17 @@
+import { defineMessages } from "react-intl";
+
+const messages = defineMessages({
+  azureFunctionsOriginalTitle: {
+    id: "azureFunctions.originalTitle",
+    defaultMessage: "Azure Functions"
+  },
+  cosmosOriginalTitle: {
+    id: "cosmosDb.originalTitle",
+    defaultMessage: "CosmosDB"
+  },
+  appServiceOriginalTitle: {
+    id: "appService.originalTitle",
+    defaultMessage: "App Service"
+  }
+});
+export default messages;
